+++ src/BulkWriter/Pipelines/ICaster.cs
@@ -1,0 +1,7 @@
+﻿namespace BulkWriter.Pipelines
+{
+    public interface ICaster<in TIn, out TOut>
+    {
+        TOut Cast(TIn input);
+    }
+}
