+++ src/BulkWriter/Pipelines/IDecorator.cs
@@ -1,0 +1,7 @@
+﻿namespace BulkWriter.Pipelines
+{
+    public interface IDecorator<in TInput, TOutput>
+    {
+        TOutput Decorate(TInput input, TOutput target);
+    }
+}
