+++ src/BulkWriter/Pipelines/IGenerator.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace BulkWriter.Pipelines
+{
+    public interface IGenerator<out T>
+    {
+        IEnumerable<T> Select();
+    }
+}
