+++ src/BulkWriter/Pipelines/IMapper.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace BulkWriter.Pipelines
+{
+    public interface IMapper<in TInput, out TOutput>
+    {
+        IEnumerable<TOutput> Map(TInput input);
+    }
+}
