+++ src/BulkWriter/Pipelines/IReducer.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace BulkWriter.Pipelines
+{
+    public interface IReducer<in TInput, out TOutput>
+    {
+        IEnumerable<TOutput> Reduce(IEnumerable<TInput> input);
+    }
+}
